ai-services/requirements-ai.txt
# Core AI/ML dependencies
transformers==4.35.0
torch==2.1.0
torchaudio==2.1.0
openai-whisper==20231117
sentencepiece==0.1.99
accelerate==0.25.0

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangChain for LLM orchestration
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20

# Audio processing
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# Scientific computing
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
